Reapply "vercel-queue: Remove `asgi_app()` (#309)" (#324)
This reverts commit 65b2c09cab006c2cb22718aeb8a4ff5817c395fb.
