--- changes/vercel-queue/20260824224130.internal.md
@@ -1,1 +1,0 @@
-Temporarily revert an unreleased change.
