+++ changes/vercel-queue/remove-asgi-app.breaking.md
@@ -1,0 +1,3 @@
+Remove the `asgi_app`, `QueueClientAsgiApp`, `QueueClient.asgi_app()`,
+`QueueClientAsgiDevServer`, and `queue_client_asgi_dev_server` APIs. Deploy
+queue handlers with `[[tool.vercel.subscribers]]` declarations instead.
