+++ src/vercel-queue/benchmarks/async_load.py
@@ -1,48 +1,33 @@
 from __future__ import annotations
 
 from typing import Any, Literal, cast
-from typing_extensions import Self
 
 import argparse
 import cProfile
 import json
-import os
 import pstats
-import subprocess  # noqa: S404
 import sys
 import threading
 import time
-from collections.abc import AsyncIterable, AsyncIterator, Callable, Mapping, Sequence
-from contextlib import AsyncExitStack
+from collections.abc import AsyncIterable, AsyncIterator, Sequence
 from dataclasses import dataclass, field
 from pathlib import Path
-from types import TracebackType
 
 import anyio
-import httpx
-from anyio import to_thread
-from anyio.abc import Process
 from anyio.lowlevel import checkpoint
 
 from vercel.queue import (
-    ALL_DEPLOYMENTS,
     ByteStreamTransport,
     Delivery,
     QueueClient,
     Topic,
     subscribe,
 )
-from vercel.queue._internal import lease as lease_internal  # noqa: PLC2701
-from vercel.queue._internal.embedded import EmbeddedQueueDispatcher  # noqa: PLC2701
-from vercel.queue._internal.lease import reset_lease_renewal_worker_for_tests  # noqa: PLC2701
-from vercel.queue.devserver import embedded_queue_dev_server, queue_client_asgi_dev_server
 from vercel.queue.embedded import embedded_queue_service
 from vercel.queue.testing import clear_subscriptions
 
 ScenarioName = Literal["small", "large"]
 DeliveryMode = Literal["pull", "push"]
-Architecture = Literal["in-process", "split"]
-ClientName = Literal["python", "js"]
 DEFAULT_SMALL_MESSAGES = 10_000
 DEFAULT_LARGE_MESSAGES = 100
 DEFAULT_CONCURRENCY = 100
@@ -53,18 +38,6 @@ DEFAULT_CHUNK_BYTES = 64 * 1024
 TOPIC = "bench-topic"
 CONSUMER_GROUP = "bench-consumer"
 LOCAL_TOKEN = "local-token"  # noqa: S105
-MODULE_PATH = str(Path(__file__).resolve())
-PROJECT_ROOT = Path(__file__).resolve().parents[3]
-BENCHMARK_ROOT = PROJECT_ROOT / ".benchmarks"
-JS_CLIENT_ROOT = BENCHMARK_ROOT / "js-client"
-JS_CLIENT_SCRIPT = JS_CLIENT_ROOT / "client.mjs"
-ROLE_NEWLINE = b"\n"
-ENV_QUEUE_SERVER_PROFILE = "VERCEL_QUEUE_BENCH_QUEUE_SERVER_PROFILE"
-ENV_CLIENT_SERVER_PROFILE = "VERCEL_QUEUE_BENCH_CLIENT_SERVER_PROFILE"
-ENV_LEASE_WORKER_PROFILE = "VERCEL_QUEUE_BENCH_LEASE_WORKER_PROFILE"
-ENV_SPLIT_ROLE = "VERCEL_QUEUE_BENCH_ROLE"
-ENV_QUEUE_BASE_URL = "VERCEL_QUEUE_BENCH_QUEUE_BASE_URL"
-ENV_BENCH_CONFIG = "VERCEL_QUEUE_BENCH_CONFIG"
 
 
 @dataclass(frozen=True)
@@ -77,12 +50,6 @@ class BenchmarkConfig:
     small_payload_bytes: int = DEFAULT_SMALL_PAYLOAD_BYTES
     large_payload_bytes: int = DEFAULT_LARGE_PAYLOAD_BYTES
     chunk_size: int = DEFAULT_CHUNK_BYTES
-    architecture: Architecture = "in-process"
-    client: ClientName = "python"
-
-    def __post_init__(self) -> None:
-        if self.client == "js" and (self.architecture != "split" or self.delivery != "push"):
-            raise ValueError("--client js requires --architecture split --delivery push")
 
 
 @dataclass(frozen=True)
@@ -171,8 +138,6 @@ class BenchmarkResult:
                 "small_payload_bytes": self.config.small_payload_bytes,
                 "large_payload_bytes": self.config.large_payload_bytes,
                 "chunk_size": self.config.chunk_size,
-                "architecture": self.config.architecture,
-                "client": self.config.client,
             },
             "total_seconds": self.total_seconds,
             "messages_per_second": self.messages_per_second,
@@ -219,7 +184,7 @@ class _PushState:
 
 
 @dataclass
-class _SplitCallbackMetrics:
+class _CallbackMetrics:
     lock: threading.Lock = field(default_factory=threading.Lock)
     messages: int = 0
     errors: int = 0
@@ -237,7 +202,7 @@ class _SplitCallbackMetrics:
 
 
 class _BenchmarkPushClient(QueueClient):
-    def __init__(self, *, metrics: _SplitCallbackMetrics, **kwargs: Any) -> None:
+    def __init__(self, *, metrics: _CallbackMetrics, **kwargs: Any) -> None:
         super().__init__(**kwargs)
         self._metrics = metrics
 
@@ -283,18 +248,6 @@ def parse_args(argv: Sequence[str] | None = None) -> argparse.Namespace:
         choices=["pull", "push", "all"],
         default="pull",
         help="delivery path to benchmark",
-    )
-    parser.add_argument(
-        "--architecture",
-        choices=["in-process", "split"],
-        default="in-process",
-        help="run push delivery in-process or split across localhost uvicorn servers",
-    )
-    parser.add_argument(
-        "--client",
-        choices=["python", "js"],
-        default="python",
-        help="callback client implementation for split push benchmarks",
     )
     parser.add_argument(
         "--messages",
@@ -349,33 +302,22 @@ def parse_args(argv: Sequence[str] | None = None) -> argparse.Namespace:
         default=None,
         help="write cProfile data to this path and print top cumulative functions",
     )
-    parser.add_argument(
-        "--split-profile-dir",
-        type=Path,
-        default=None,
-        help="write separate cProfile data for split queue and client processes",
-    )
     return parser.parse_args(argv)
 
 
 async def run_benchmarks(
     configs: Sequence[BenchmarkConfig],
-    profile_dir: Path | None = None,
 ) -> list[BenchmarkResult]:
-    return [await run_benchmark(config, profile_dir=profile_dir) for config in configs]
+    return [await run_benchmark(config) for config in configs]
 
 
 async def run_benchmark(
     config: BenchmarkConfig,
-    *,
-    profile_dir: Path | None = None,
 ) -> BenchmarkResult:
     _validate_config(config)
-    if config.architecture == "split":
-        return await _run_split_benchmark(config, profile_dir=profile_dir)
     total_started = time.perf_counter()
     clear_subscriptions()
-    metrics = _SplitCallbackMetrics()
+    metrics = _CallbackMetrics()
     async with embedded_queue_service() as service:
         if config.delivery == "push":
 
@@ -438,8 +380,6 @@ def configs_from_args(args: argparse.Namespace) -> list[BenchmarkConfig]:
                     small_payload_bytes=args.small_size,
                     large_payload_bytes=args.large_size,
                     chunk_size=args.chunk_size,
-                    architecture=args.architecture,
-                    client=args.client,
                 )
             )
     return configs
@@ -450,8 +390,6 @@ def format_text(results: Sequence[BenchmarkResult]) -> str:
         "\n".join([
             f"scenario: {result.scenario}",
             f"delivery: {result.config.delivery}",
-            f"architecture: {result.config.architecture}",
-            f"client: {result.config.client}",
             f"messages: {result.config.messages}",
             (
                 f"total: {_seconds(result.total_seconds)} "
@@ -469,24 +407,16 @@ def format_text(results: Sequence[BenchmarkResult]) -> str:
 
 def main(argv: Sequence[str] | None = None) -> int:
     args = parse_args(argv)
-    role = os.environ.get(ENV_SPLIT_ROLE, "driver")
-    if role == "split-queue":
-        return _profiled_role(args.profile, lambda: _run_split_queue_role(args))
-    if role == "split-client":
-        return _profiled_role(args.profile, lambda: _run_split_client_role(args))
-    if role != "driver":
-        raise ValueError(f"unknown benchmark role: {role!r}")
     configs = configs_from_args(args)
 
     if args.profile is None:
-        results = _run_configs_sync(configs, args.backend, profile_dir=args.split_profile_dir)
+        results = _run_configs_sync(configs, args.backend)
     else:
         profiler = cProfile.Profile()
         results = profiler.runcall(
             _run_configs_sync,
             configs,
             args.backend,
-            profile_dir=args.split_profile_dir,
         )
         profiler.dump_stats(args.profile)
         stats = pstats.Stats(profiler, stream=sys.stderr).sort_stats("cumulative")
@@ -501,651 +431,11 @@ def main(argv: Sequence[str] | None = None) -> int:
     return 0
 
 
-def _profiled_role(profile: Path | None, run: Callable[[], int]) -> int:
-    if profile is None:
-        return run()
-    profiler = cProfile.Profile()
-    try:
-        return profiler.runcall(run)
-    finally:
-        profiler.dump_stats(profile)
-
-
-def _profile_lease_renewal_worker(profile: str | None) -> Callable[[], None]:
-    original = lease_internal._run_lease_renewal_worker_thread  # noqa: SLF001
-    if profile is None:
-        return lambda: None
-
-    def profiled_worker() -> None:
-        profiler = cProfile.Profile()
-        try:
-            profiler.runcall(original)
-        finally:
-            profiler.dump_stats(profile)
-
-    lease_internal._run_lease_renewal_worker_thread = profiled_worker  # noqa: SLF001  # ty: ignore[invalid-assignment]
-
-    def restore() -> None:
-        lease_internal._run_lease_renewal_worker_thread = original  # noqa: SLF001
-
-    return restore
-
-
 def _run_configs_sync(
     configs: Sequence[BenchmarkConfig],
     backend: str,
-    *,
-    profile_dir: Path | None = None,
 ) -> list[BenchmarkResult]:
-    return anyio.run(run_benchmarks, configs, profile_dir, backend=backend)
-
-
-JS_CLIENT_HARNESS = r"""
-import http from "node:http";
-import { QueueClient } from "@vercel/queue";
-
-const TOPIC = "bench-topic";
-const LOCAL_TOKEN = "local-token";
-const queueBaseUrl = process.env.VERCEL_QUEUE_BENCH_QUEUE_BASE_URL;
-const config = JSON.parse(process.env.VERCEL_QUEUE_BENCH_CONFIG ?? "{}");
-
-if (!queueBaseUrl) {
-  throw new Error("VERCEL_QUEUE_BENCH_QUEUE_BASE_URL is required");
-}
-
-const metrics = {
-  messages: 0,
-  bytes: 0,
-  errors: 0,
-  inFlight: 0,
-  peakInFlight: 0,
-  latencies: [],
-};
-
-const client = new QueueClient({
-  token: LOCAL_TOKEN,
-  region: "iad1",
-  deploymentId: null,
-  resolveBaseUrl: () => new URL(queueBaseUrl),
-});
-
-function jsonPayloadSize(payload) {
-  return Buffer.byteLength(JSON.stringify(payload));
-}
-
-async function payloadSize(payload) {
-  if (config.scenario !== "large") {
-    return jsonPayloadSize(payload);
-  }
-  if (payload instanceof Uint8Array) {
-    return payload.byteLength;
-  }
-  if (payload instanceof ArrayBuffer) {
-    return payload.byteLength;
-  }
-  if (payload && typeof payload.getReader === "function") {
-    let total = 0;
-    const reader = payload.getReader();
-    try {
-      while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        total += value?.byteLength ?? 0;
-      }
-    } finally {
-      reader.releaseLock();
-    }
-    return total;
-  }
-  return jsonPayloadSize(payload);
-}
-
-const callback = client.handleCallback(async (payload) => {
-  metrics.inFlight += 1;
-  metrics.peakInFlight = Math.max(metrics.peakInFlight, metrics.inFlight);
-  try {
-    metrics.bytes += await payloadSize(payload);
-    metrics.messages += 1;
-  } catch (error) {
-    metrics.errors += 1;
-    throw error;
-  } finally {
-    metrics.inFlight -= 1;
-  }
-});
-
-const server = http.createServer(async (req, res) => {
-  const chunks = [];
-  for await (const chunk of req) {
-    chunks.push(chunk);
-  }
-  const body = Buffer.concat(chunks);
-  const request = new Request(`http://localhost${req.url ?? "/"}`, {
-    method: req.method,
-    headers: req.headers,
-    body,
-  });
-  const started = performance.now();
-  try {
-    const response = await callback(request);
-    metrics.latencies.push((performance.now() - started) / 1000);
-    res.writeHead(response.status, Object.fromEntries(response.headers));
-    res.end(Buffer.from(await response.arrayBuffer()));
-  } catch (error) {
-    metrics.errors += 1;
-    metrics.latencies.push((performance.now() - started) / 1000);
-    res.writeHead(500, { "content-type": "application/json" });
-    res.end(JSON.stringify({ error: String(error) }));
-  }
-});
-
-function writeJson(payload) {
-  process.stdout.write(`${JSON.stringify(payload)}\n`);
-}
-
-server.listen(0, "127.0.0.1", () => {
-  const address = server.address();
-  writeJson({ baseUrl: `http://127.0.0.1:${address.port}` });
-});
-
-process.stdin.setEncoding("utf8");
-let buffer = "";
-process.stdin.on("data", (chunk) => {
-  buffer += chunk;
-  let newline;
-  while ((newline = buffer.indexOf("\n")) >= 0) {
-    const line = buffer.slice(0, newline);
-    buffer = buffer.slice(newline + 1);
-    if (!line) continue;
-    const command = JSON.parse(line);
-    if (command.command === "status") {
-      writeJson({
-        messages: metrics.messages,
-        bytes: metrics.bytes,
-        errors: metrics.errors,
-        peak_in_flight: metrics.peakInFlight,
-        latencies: [],
-        callback_messages: metrics.messages,
-        callback_latencies: metrics.latencies,
-      });
-    } else if (command.command === "stop") {
-      writeJson({ ok: true });
-      server.close(() => process.exit(0));
-    } else {
-      writeJson({ error: `unknown command: ${command.command}` });
-    }
-  }
-});
-"""
-
-
-async def _prepare_js_client_harness() -> Path:
-    await to_thread.run_sync(_prepare_js_client_harness_sync)
-    return JS_CLIENT_SCRIPT
-
-
-def _prepare_js_client_harness_sync() -> None:
-    JS_CLIENT_ROOT.mkdir(parents=True, exist_ok=True)
-    package_json = JS_CLIENT_ROOT / "package.json"
-    if not package_json.exists():
-        package_json.write_text(
-            json.dumps(
-                {
-                    "private": True,
-                    "type": "module",
-                    "dependencies": {"@vercel/queue": "latest"},
-                },
-                indent=2,
-            )
-            + "\n"
-        )
-    JS_CLIENT_SCRIPT.write_text(JS_CLIENT_HARNESS.lstrip() + "\n")
-    node_modules = JS_CLIENT_ROOT / "node_modules" / "@vercel" / "queue"
-    if node_modules.exists():
-        return
-    npm = os.environ.get("NPM", "npm")
-    subprocess.run(  # noqa: S603
-        [npm, "install", "--silent", "--no-audit", "--no-fund"],
-        cwd=JS_CLIENT_ROOT,
-        check=True,
-    )
-
-
-class _HttpPushCallbackDispatcher:
-    def __init__(self, base_url: str) -> None:
-        self._base_url = base_url
-        self._client = httpx.AsyncClient(timeout=30)
-
-    async def __aenter__(self) -> Self:
-        return self
-
-    async def __aexit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc: BaseException | None,
-        traceback: TracebackType | None,
-    ) -> None:
-        del exc_type, exc, traceback
-        await self.aclose()
-
-    async def accept_and_handle(self, raw_body: bytes, headers: Mapping[str, str]) -> None:
-        response = await self._client.post(self._base_url, content=raw_body, headers=headers)
-        response.raise_for_status()
-
-    async def aclose(self) -> None:
-        await self._client.aclose()
-
-
-async def _run_split_benchmark(
-    config: BenchmarkConfig,
-    *,
-    profile_dir: Path | None = None,
-) -> BenchmarkResult:
-    if config.delivery != "push":
-        raise ValueError("split architecture currently supports push delivery only")
-    total_started = time.perf_counter()
-    clear_subscriptions()
-    split_processes = await _SplitProcesses.open(config, profile_dir=profile_dir)
-    async with split_processes as processes:
-        client = QueueClient(
-            token=LOCAL_TOKEN,
-            region="iad1",
-            base_url=processes.queue_base_url,
-            deployment=ALL_DEPLOYMENTS,
-        )
-        send = await _send_messages(client, config)
-        receive_ack = await _wait_for_split_push_messages(
-            processes,
-            config,
-            timeout_seconds=300.0 if profile_dir is not None else 30.0,
-        )
-    total_seconds = time.perf_counter() - total_started
-    try:
-        return BenchmarkResult(
-            scenario=config.scenario,
-            config=config,
-            total_seconds=total_seconds,
-            send=send.phase,
-            receive_ack=receive_ack.phase,
-            errors=send.errors
-            + receive_ack.errors
-            + max(0, config.messages - receive_ack.phase.messages),
-            peak_in_flight=max(send.peak_in_flight, receive_ack.peak_in_flight),
-        )
-    finally:
-        clear_subscriptions()
-
-
-@dataclass
-class _SplitProcesses:
-    queue: Process
-    client: Process
-    queue_base_url: str
-
-    @classmethod
-    async def open(
-        cls,
-        config: BenchmarkConfig,
-        *,
-        profile_dir: Path | None,
-    ) -> _SplitProcessesContext:
-        context = _SplitProcessesContext(config, profile_dir=profile_dir)
-        await context.start()
-        return context
-
-    async def request_queue(self, payload: dict[str, Any]) -> dict[str, Any]:
-        return await _process_request(self.queue, payload)
-
-    async def request_client(self, payload: dict[str, Any]) -> dict[str, Any]:
-        return await _process_request(self.client, payload)
-
-
-class _SplitProcessesContext(_SplitProcesses):
-    def __init__(self, config: BenchmarkConfig, *, profile_dir: Path | None) -> None:
-        self._config = config
-        self._profile_dir = profile_dir
-        self.queue_base_url = ""
-
-    async def start(self) -> None:
-        profile_dir = self._profile_dir
-        if profile_dir is not None:
-            profile_dir.mkdir(parents=True, exist_ok=True)
-        self.queue = await _start_role_process(
-            self._config,
-            "split-queue",
-            profile=(profile_dir / "queue.prof" if profile_dir is not None else None),
-            queue_server_profile=(
-                profile_dir / "queue-server.prof" if profile_dir is not None else None
-            ),
-        )
-        queue_ready = await _read_process_json(self.queue)
-        if self._config.client == "js":
-            self.client = await _start_js_client_process(
-                self._config,
-                queue_base_url=str(queue_ready["baseUrl"]),
-            )
-        else:
-            self.client = await _start_role_process(
-                self._config,
-                "split-client",
-                queue_base_url=queue_ready["baseUrl"],
-                profile=(profile_dir / "client.prof" if profile_dir is not None else None),
-                client_server_profile=(
-                    profile_dir / "client-server.prof" if profile_dir is not None else None
-                ),
-                lease_worker_profile=(
-                    profile_dir / "client-lease-worker.prof" if profile_dir is not None else None
-                ),
-            )
-        client_ready = await _read_process_json(self.client)
-        self.queue_base_url = str(queue_ready["baseUrl"])
-        await self.request_queue({"command": "set_callback", "baseUrl": client_ready["baseUrl"]})
-
-    async def __aenter__(self) -> Self:
-        return self
-
-    async def __aexit__(self, *args: object) -> None:
-        for proc in (self.queue, self.client):
-            try:
-                await _process_request(proc, {"command": "stop"})
-            except RuntimeError:
-                proc.terminate()
-        with anyio.move_on_after(5):
-            await self.queue.wait()
-            await self.client.wait()
-        for proc in (self.queue, self.client):
-            if proc.returncode is None:
-                proc.kill()
-                await proc.wait()
-
-
-async def _start_role_process(
-    config: BenchmarkConfig,
-    role: str,
-    *,
-    queue_base_url: str | None = None,
-    profile: Path | None = None,
-    queue_server_profile: Path | None = None,
-    client_server_profile: Path | None = None,
-    lease_worker_profile: Path | None = None,
-) -> Process:
-    env = os.environ.copy()
-    command = [
-        sys.executable,
-        MODULE_PATH,
-        "--scenario",
-        config.scenario,
-        "--delivery",
-        config.delivery,
-        "--architecture",
-        config.architecture,
-        "--client",
-        config.client,
-        "--messages",
-        str(config.messages),
-        "--concurrency",
-        str(config.concurrency),
-        "--small-size",
-        str(config.small_payload_bytes),
-        "--large-size",
-        str(config.large_payload_bytes),
-        "--chunk-size",
-        str(config.chunk_size),
-    ]
-    env[ENV_SPLIT_ROLE] = role
-    if queue_base_url is not None:
-        env[ENV_QUEUE_BASE_URL] = queue_base_url
-    if profile is not None:
-        command.extend(["--profile", str(profile)])
-    if queue_server_profile is not None:
-        env[ENV_QUEUE_SERVER_PROFILE] = str(queue_server_profile)
-    if client_server_profile is not None:
-        env[ENV_CLIENT_SERVER_PROFILE] = str(client_server_profile)
-    if lease_worker_profile is not None:
-        env[ENV_LEASE_WORKER_PROFILE] = str(lease_worker_profile)
-    return await anyio.open_process(command, env=env)
-
-
-async def _start_js_client_process(
-    config: BenchmarkConfig,
-    *,
-    queue_base_url: str,
-) -> Process:
-    script = await _prepare_js_client_harness()
-    env = os.environ.copy()
-    env[ENV_QUEUE_BASE_URL] = queue_base_url
-    env[ENV_BENCH_CONFIG] = _config_env(config)
-    return await anyio.open_process(["node", str(script)], cwd=JS_CLIENT_ROOT, env=env)
-
-
-async def _read_process_json(process: Process) -> dict[str, Any]:
-    stdout = process.stdout
-    if stdout is None:
-        raise RuntimeError("role process stdout is unavailable")
-    line = bytearray()
-    while True:
-        chunk = await stdout.receive(1)
-        if not chunk:
-            raise RuntimeError("role process closed stdout before sending JSON")
-        if chunk == ROLE_NEWLINE:
-            break
-        line.extend(chunk)
-    return cast("dict[str, Any]", json.loads(line.decode()))
-
-
-async def _process_request(
-    process: Process,
-    payload: dict[str, Any],
-) -> dict[str, Any]:
-    stdin = process.stdin
-    if stdin is None:
-        raise RuntimeError("role process stdin is unavailable")
-    await stdin.send(json.dumps(payload).encode() + ROLE_NEWLINE)
-    return await _read_process_json(process)
-
-
-async def _wait_for_split_push_messages(
-    processes: _SplitProcesses,
-    config: BenchmarkConfig,
-    *,
-    timeout_seconds: float = 30.0,
-) -> _ReceiveResult:
-    started = time.perf_counter()
-    deadline = started + timeout_seconds
-    client_status: dict[str, Any] = {}
-    queue_status: dict[str, Any] = {}
-    while time.perf_counter() < deadline:
-        client_status = await processes.request_client({"command": "status"})
-        queue_status = await processes.request_queue({"command": "status"})
-        if (
-            client_status["callback_messages"] >= config.messages
-            and queue_status["acknowledged"] >= config.messages
-        ):
-            break
-        await anyio.sleep(0.01)
-    elapsed = time.perf_counter() - started
-    latencies = cast("list[float]", client_status.get("callback_latencies", []))
-    if not latencies:
-        latencies = cast("list[float]", client_status.get("latencies", []))
-    return _ReceiveResult(
-        phase=PhaseResult(
-            seconds=elapsed,
-            messages=int(client_status.get("messages", 0)),
-            bytes=int(client_status.get("bytes", 0)),
-            latency=LatencySummary.from_samples(latencies),
-        ),
-        errors=int(client_status.get("errors", 0)) + int(queue_status.get("errors", 0)),
-        peak_in_flight=max(
-            int(client_status.get("peak_in_flight", 0)),
-            int(queue_status.get("peak_in_flight", 0)),
-        ),
-    )
-
-
-def _run_split_queue_role(args: argparse.Namespace) -> int:
-    anyio.run(_run_split_queue_role_async, args)
-    return 0
-
-
-async def _run_split_queue_role_async(args: argparse.Namespace) -> None:
-    config = configs_from_args(args)[0]
-    callback_dispatcher: _HttpPushCallbackDispatcher | None = None
-    async with anyio.create_task_group() as task_group, AsyncExitStack() as exit_stack:
-        with embedded_queue_dev_server(
-            manual_clock=False,
-            profile=os.environ.get(ENV_QUEUE_SERVER_PROFILE),
-        ) as queue_server:
-
-            def callback_dispatcher_factory() -> _HttpPushCallbackDispatcher:
-                if callback_dispatcher is None:
-                    raise RuntimeError("split callback dispatcher is not registered")
-                return callback_dispatcher
-
-            dispatcher = EmbeddedQueueDispatcher(
-                queue_server.app.server,
-                callback_dispatcher_factory,
-            )
-            task_group.start_soon(dispatcher.run, name="vercel-split-queue-dispatcher")
-            await _write_role_json({"baseUrl": queue_server.base_url})
-            try:
-                while True:
-                    command = await _read_role_command()
-                    if command["command"] == "set_callback":
-                        callback_dispatcher = await exit_stack.enter_async_context(
-                            _HttpPushCallbackDispatcher(str(command["baseUrl"]))
-                        )
-                        dispatcher.register_subscription(
-                            topic=TOPIC,
-                            consumer_group=CONSUMER_GROUP,
-                            retry_after_seconds=None,
-                            initial_delay_seconds=None,
-                            max_concurrency=config.concurrency,
-                            max_attempts=None,
-                        )
-                        dispatcher.wake()
-                        await _write_role_json({"ok": True})
-                    elif command["command"] == "status":
-                        await _write_role_json(_split_queue_status(queue_server, dispatcher))
-                    elif command["command"] == "stop":
-                        await _write_role_json({"ok": True})
-                        return
-                    else:
-                        await _write_role_json({"error": f"unknown command: {command['command']}"})
-            finally:
-                task_group.cancel_scope.cancel()
-                dispatcher.unregister()
-                await dispatcher.aclose()
-
-
-def _run_split_client_role(args: argparse.Namespace) -> int:
-    anyio.run(_run_split_client_role_async, args)
-    return 0
-
-
-async def _run_split_client_role_async(args: argparse.Namespace) -> None:
-    queue_base_url = os.environ.get(ENV_QUEUE_BASE_URL)
-    if queue_base_url is None:
-        raise ValueError(f"split client role requires {ENV_QUEUE_BASE_URL}")
-    config = configs_from_args(args)[0]
-    state = _PushState()
-    metrics = _SplitCallbackMetrics()
-    _register_split_client_handler(config, state)
-    restore_lease_worker = _profile_lease_renewal_worker(os.environ.get(ENV_LEASE_WORKER_PROFILE))
-    client = _BenchmarkPushClient(
-        metrics=metrics,
-        token=LOCAL_TOKEN,
-        region="iad1",
-        base_url=queue_base_url,
-        deployment=ALL_DEPLOYMENTS,
-    )
-    try:
-        with queue_client_asgi_dev_server(
-            client=client,
-            profile=os.environ.get(ENV_CLIENT_SERVER_PROFILE),
-        ) as server:
-            await _write_role_json({"baseUrl": server.base_url})
-            while True:
-                command = await _read_role_command()
-                if command["command"] == "status":
-                    await _write_role_json(_split_client_status(state, metrics))
-                elif command["command"] == "stop":
-                    await _write_role_json({"ok": True})
-                    return
-                else:
-                    await _write_role_json({"error": f"unknown command: {command['command']}"})
-    finally:
-        reset_lease_renewal_worker_for_tests()
-        restore_lease_worker()
-
-
-def _register_split_client_handler(
-    config: BenchmarkConfig,
-    state: _PushState,
-) -> None:
-    if config.scenario == "large":
-
-        @subscribe(topic=TOPIC, consumer_group=CONSUMER_GROUP, max_concurrency=config.concurrency)
-        async def handle_large(payload: AsyncIterable[bytes]) -> None:
-            await _handle_push_payload(state, config, payload)
-
-        state.handlers.append(handle_large)
-    else:
-
-        @subscribe(topic=TOPIC, consumer_group=CONSUMER_GROUP, max_concurrency=config.concurrency)
-        async def handle_small(payload: dict[str, Any]) -> None:
-            await _handle_push_payload(state, config, payload)
-
-        state.handlers.append(handle_small)
-
-
-def _require_callback_base_url(base_url: str | None) -> str:
-    if base_url is None:
-        raise RuntimeError("split callback server is not registered")
-    return base_url
-
-
-def _split_queue_status(queue_server: Any, dispatcher: EmbeddedQueueDispatcher) -> dict[str, Any]:
-    messages = [
-        message
-        for message in queue_server.app.server.state.messages
-        if message.topic == TOPIC and message.payload != {"warm": True}
-    ]
-    acknowledged = sum(1 for message in messages if message.acknowledged_for(CONSUMER_GROUP))
-    return {
-        "messages": len(messages),
-        "acknowledged": acknowledged,
-        "errors": 0,
-        "peak_in_flight": max(dispatcher._inflight_counts.values(), default=0),  # noqa: SLF001
-    }
-
-
-def _split_client_status(
-    state: _PushState,
-    metrics: _SplitCallbackMetrics,
-) -> dict[str, Any]:
-    with metrics.lock:
-        callback_messages = metrics.messages
-        callback_errors = metrics.errors
-        callback_latencies = list(metrics.latencies)
-    return {
-        "messages": state.counters.messages,
-        "bytes": state.counters.bytes,
-        "errors": state.counters.errors + callback_errors,
-        "peak_in_flight": state.peak_in_flight,
-        "latencies": state.counters.latencies,
-        "callback_messages": callback_messages,
-        "callback_latencies": callback_latencies,
-    }
-
-
-async def _read_role_command() -> dict[str, Any]:
-    return cast("dict[str, Any]", json.loads(await to_thread.run_sync(sys.stdin.readline)))
-
-
-async def _write_role_json(payload: dict[str, Any]) -> None:
-    line = json.dumps(payload, separators=(",", ":")) + "\n"
-    await to_thread.run_sync(sys.stdout.write, line)
-    await to_thread.run_sync(sys.stdout.flush)
+    return anyio.run(run_benchmarks, configs, backend=backend)
 
 
 @dataclass(frozen=True)
@@ -1348,7 +638,7 @@ async def _wait_for_push_messages(
     service: Any,
     receive_ack: _ReceiveResult,
     config: BenchmarkConfig,
-    metrics: _SplitCallbackMetrics,
+    metrics: _CallbackMetrics,
 ) -> _ReceiveResult:
     if receive_ack.state is None or receive_ack.started is None:
         raise RuntimeError("push receive state was not initialized")
@@ -1374,12 +664,12 @@ async def _wait_for_push_messages(
     return receive_ack
 
 
-def _callback_metrics(metrics: _SplitCallbackMetrics) -> tuple[int, int, list[float]]:
+def _callback_metrics(metrics: _CallbackMetrics) -> tuple[int, int, list[float]]:
     with metrics.lock:
         return metrics.messages, metrics.errors, list(metrics.latencies)
 
 
-def _callback_message_count(metrics: _SplitCallbackMetrics) -> int:
+def _callback_message_count(metrics: _CallbackMetrics) -> int:
     with metrics.lock:
         return metrics.messages
 
@@ -1449,26 +739,6 @@ def _validate_config(config: BenchmarkConfig) -> None:
     _positive_int(config.small_payload_bytes)
     _positive_int(config.large_payload_bytes)
     _positive_int(config.chunk_size)
-    if config.client == "js" and (config.architecture != "split" or config.delivery != "push"):
-        raise ValueError("--client js requires --architecture split --delivery push")
-
-
-def _config_env(config: BenchmarkConfig) -> str:
-    return json.dumps(
-        {
-            "scenario": config.scenario,
-            "delivery": config.delivery,
-            "messages": config.messages,
-            "concurrency": config.concurrency,
-            "poll_limit": config.poll_limit,
-            "small_payload_bytes": config.small_payload_bytes,
-            "large_payload_bytes": config.large_payload_bytes,
-            "chunk_size": config.chunk_size,
-            "architecture": config.architecture,
-            "client": config.client,
-        },
-        separators=(",", ":"),
-    )
 
 
 def _positive_int(value: str | int) -> int:
