+++ src/vercel-queue/examples/binary_buffer/main.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
-from vercel.queue import Message, asgi_app, subscribe
+from vercel.queue import Message, subscribe
 
 
 @subscribe(topic="images")
 def handle_image(message: Message[bytes]) -> None:
     print("Received image bytes", message.payload.hex())
-
-
-app = asgi_app()
