+++ src/vercel-queue/examples/custom_transport/main.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from invoice_transport import Invoice, InvoiceFormTransport
 
-from vercel.queue import Topic, asgi_app, subscribe
+from vercel.queue import Topic, subscribe
 
 invoice_events = Topic[Invoice]("invoices", transport=InvoiceFormTransport())
 
@@ -10,6 +10,3 @@ invoice_events = Topic[Invoice]("invoices", transport=InvoiceFormTransport())
 @subscribe(topic=invoice_events)
 async def record_invoice(invoice: Invoice) -> None:
     print("Recorded invoice", invoice.invoice_id, invoice.customer_id, invoice.total_cents)
-
-
-app = asgi_app()
