+++ src/vercel-queue/examples/large_file_stream/main.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from collections.abc import AsyncIterable
 
-from vercel.queue import asgi_app, subscribe
+from vercel.queue import subscribe
 
 
 @subscribe(topic="files")
@@ -11,6 +11,3 @@ async def archive_file(payload: AsyncIterable[bytes]) -> None:
     async for chunk in payload:
         size += len(chunk)
     print(f"received {size} bytes")
-
-
-app = asgi_app()
