+++ src/vercel-queue/examples/text_stream/main.py
@@ -2,12 +2,9 @@ from __future__ import annotations
 
 from collections.abc import Iterable
 
-from vercel.queue import asgi_app, subscribe
+from vercel.queue import subscribe
 
 
 @subscribe(topic="logs")
 def index_logs(payload: Iterable[str]) -> None:
     print("".join(payload), end="")
-
-
-app = asgi_app()
