+++ src/vercel-queue/vercel/queue/_internal/devserver.py
@@ -12,35 +12,10 @@ import json
 import signal
 import threading
 import time
-from collections.abc import Callable, Iterator, Mapping
-from dataclasses import dataclass
-
-import anyio
-from anyio import from_thread
+from collections.abc import Callable, Iterator
 
 from ..embedded import create_embedded_queue_app
-from .asgi import QueueClientAsgiApp, asgi_app
-from .client import QueueClient
-from .config import CURRENT_DEPLOYMENT, BaseUrl, DeploymentOption
 from .embedded import EmbeddedQueueDevServer
-from .http import AsyncHttpClientFactory
-from .types import Duration
-
-_GRACEFUL_SHUTDOWN_TIMEOUT_SECONDS = 5.0
-_SERVER_STOP_TIMEOUT_MARGIN_SECONDS = 1.0
-
-
-@dataclass(frozen=True, kw_only=True)
-class QueueClientAsgiDevServer:
-    """Running queue client ASGI dev server."""
-
-    base_url: str
-    app: QueueClientAsgiApp
-    _server: Any
-    _thread: threading.Thread
-
-    def is_running(self) -> bool:
-        return self._thread.is_alive() and not self._server.should_exit
 
 
 @contextlib.contextmanager
@@ -61,9 +36,8 @@ def embedded_queue_dev_server(
         log_level="warning",
         lifespan="off",
         ws="none",
-        timeout_graceful_shutdown=_GRACEFUL_SHUTDOWN_TIMEOUT_SECONDS,
     )
-    server = _cancellable_server(uvicorn, config)
+    server = uvicorn.Server(config)
     thread = threading.Thread(target=_profiled_server_run(server, profile), daemon=True)
     thread.start()
     _wait_for_server(server)
@@ -76,71 +50,11 @@ def embedded_queue_dev_server(
             _thread=thread,
         )
     finally:
-        try:
-            _stop_server(
-                server,
-                thread,
-                "embedded queue dev server",
-                timeout=_GRACEFUL_SHUTDOWN_TIMEOUT_SECONDS,
-            )
-        finally:
-            app.state.close()
-
-
-@contextlib.contextmanager
-def queue_client_asgi_dev_server(  # noqa: PLR0913
-    *,
-    client: QueueClient | None = None,
-    host: str = "127.0.0.1",
-    port: int = 0,
-    token: str | None = None,
-    region: str | None = None,
-    base_url: BaseUrl | None = None,
-    deployment: DeploymentOption = CURRENT_DEPLOYMENT,
-    headers: Mapping[str, str] | None = None,
-    timeout: Duration | None = 10.0,
-    http_client_factory: AsyncHttpClientFactory | None = None,
-    profile: str | None = None,
-) -> Iterator[QueueClientAsgiDevServer]:
-    """Run a queue client ASGI callback app on a localhost HTTP port."""
-    uvicorn = _uvicorn()
-    app = asgi_app(
-        client=client,
-        token=token,
-        region=region,
-        base_url=base_url,
-        deployment=deployment,
-        headers=headers,
-        timeout=timeout,
-        http_client_factory=http_client_factory,
-    )
-    config = uvicorn.Config(
-        app,
-        host=host,
-        port=port,
-        log_level="warning",
-        lifespan="on",
-        ws="none",
-        timeout_graceful_shutdown=_GRACEFUL_SHUTDOWN_TIMEOUT_SECONDS,
-    )
-    server = _cancellable_server(uvicorn, config)
-    thread = threading.Thread(target=_profiled_server_run(server, profile), daemon=True)
-    thread.start()
-    _wait_for_server(server, "queue client ASGI dev server")
-    try:
-        yield QueueClientAsgiDevServer(
-            base_url=f"http://{host}:{_server_port(server)}",
-            app=app,
-            _server=server,
-            _thread=thread,
-        )
-    finally:
-        _stop_server(
-            server,
-            thread,
-            "queue client ASGI dev server",
-            timeout=_GRACEFUL_SHUTDOWN_TIMEOUT_SECONDS,
-        )
+        server.should_exit = True
+        thread.join(timeout=5)
+        if thread.is_alive():
+            raise RuntimeError("embedded queue dev server did not stop")
+        app.state.close()
 
 
 def main(argv: list[str] | None = None) -> int:
@@ -193,52 +107,6 @@ def _wait_for_server(server: Any, name: str = "embedded queue dev server") -> No
             time.sleep(0.01)
             continue
         raise RuntimeError(f"{name} did not start")
-
-
-def _stop_server(
-    server: Any,
-    thread: threading.Thread,
-    name: str,
-    *,
-    timeout: float,
-) -> None:
-    server.should_exit = True
-    thread.join(timeout=timeout + _SERVER_STOP_TIMEOUT_MARGIN_SECONDS)
-    if thread.is_alive():
-        server.force_exit = True
-        server.cancel()
-        thread.join(timeout=timeout)
-    if thread.is_alive():
-        raise RuntimeError(f"{name} did not stop")
-
-
-def _cancellable_server(uvicorn: Any, config: Any) -> Any:
-    class CancellableServer(uvicorn.Server):
-        _cancel_scope: anyio.CancelScope | None = None
-        _portal: from_thread.BlockingPortal | None = None
-
-        def run(self, sockets: list[Any] | None = None) -> None:
-            backend_options: dict[str, Any] = {}
-            get_loop_factory = getattr(self.config, "get_loop_factory", None)
-            if get_loop_factory is not None:
-                backend_options["loop_factory"] = get_loop_factory()
-            else:
-                self.config.setup_event_loop()
-            anyio.run(self._run, sockets, backend_options=backend_options)
-
-        async def _run(self, sockets: list[Any] | None) -> None:
-            async with from_thread.BlockingPortal() as self._portal:
-                with anyio.CancelScope() as self._cancel_scope:
-                    await self.serve(sockets=sockets)
-
-        def cancel(self) -> None:
-            if self._portal is not None and self._cancel_scope is not None:
-                try:
-                    self._portal.start_task_soon(self._cancel_scope.cancel)
-                except RuntimeError:
-                    pass
-
-    return CancellableServer(config)
 
 
 def _server_port(server: Any) -> int:
