+++ src/vercel-queue/vercel/queue/devserver.py
@@ -4,19 +4,15 @@ from __future__ import annotations
 
 from ._internal.devserver import (
     EmbeddedQueueDevServer,
-    QueueClientAsgiDevServer,
     embedded_queue_dev_server,
     main,
-    queue_client_asgi_dev_server,
 )
 
 # Only add public symbols to __all__; internal helpers must stay unexported.
 __all__ = (
     "EmbeddedQueueDevServer",
-    "QueueClientAsgiDevServer",
     "embedded_queue_dev_server",
     "main",
-    "queue_client_asgi_dev_server",
 )
