+++ src/vercel-workflow/tests/unit/test_workflow_queue_transport.py
@@ -78,9 +78,9 @@ async def test_deserialize_reads_json_when_the_content_type_says_so() -> None:
 def test_dispatch_resolves_the_transport_without_help_from_the_client(
     isolated_subscriptions: None,
 ) -> None:
-    """The deployed entrypoint is a generated ``vercel.queue.asgi_app()``.
+    """The deployed subscriber entrypoint is generated by the platform.
 
-    Its client is built by the platform and carries no transport, so what
+    Its queue client carries no transport, so what
     matters is the transport dispatch resolves from the *subscription* — not
     anything hanging off a client the world constructed. Asserting the latter
     is what hid this: the world's own client never receives a delivery.
